requirements.txt
-e .
aws_cdk.aws_lambda
aws_cdk.aws_sqs
aws_cdk.aws_sns
aws_cdk.aws_cloudwatch
aws_cdk.aws_s3_notifications
aws_cdk.aws_cloudwatch_actions
aws_cdk.aws_sns_subscriptions
aws_cdk.aws_lambda_event_sources